dashboard/requirements.txt
streamlit==1.28.2
pandas==2.1.4
plotly==5.17.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0